add roles secure

+++ main.py
@@ -12,7 +12,6 @@ from flask_cors import CORS
 # pip install flask flask_restful flask_cors pymongo selenium
 
 app = Flask(__name__)
-CORS(app)
 jwt = JWTManager(app)
 app.config['JWT_SECRET_KEY'] = 'Your_Secret_Key'
 app.config['JWT_ACCESS_TOKEN_EXPIRES'] = datetime.timedelta(days=1)
@@ -24,7 +23,7 @@ def register():
     new_user = request.get_json()
     username = new_user["username"]
     password = hashlib.sha256(new_user["password"].encode("utf-8")).hexdigest()
-    role = 'ANALYST'
+    role = 'USER'
     doc = findUserByUserName(new_user["username"])
     if not doc:
         addUser(username, password, role)
@@ -40,9 +39,9 @@ def login():
     if user_from_db:
         encrpted_password = hashlib.sha256(login_details['password'].encode("utf-8")).hexdigest()
         if encrpted_password == user_from_db['password']:
-            access_token = create_access_token(identity=user_from_db['username'])
-            refresh_token = create_refresh_token(identity=user_from_db['username'])
-            return jsonify(access_token=access_token, refresh_token=refresh_token), 200
+            access_token = create_access_token(identity={'username': user_from_db['username'], 'role': user_from_db['role']})
+            # refresh_token = create_refresh_token(identity=user_from_db['username'])
+            return jsonify(access_token=access_token), 200
 
     return jsonify({'msg': 'The username or password is incorrect'}), 401
 
@@ -51,7 +50,7 @@ def login():
 @jwt_required(fresh=False)
 def profile():
     current_user = get_jwt_identity()
-    user_from_db = findUserByUserName(current_user)
+    user_from_db = findUserByUserName(current_user['username'])
     if user_from_db:
         del user_from_db['_id'], user_from_db['password']
         return jsonify({'profile': user_from_db}), 200
@@ -81,8 +80,12 @@ def getResources():
 @app.route("/api/dev-by", methods=["POST"])
 @jwt_required(fresh=False)
 def addResource():
-    addItem(request.json['date'], request.json['file'])
-    return jsonify({'msg': 'success'}), 200
+    user_role = get_jwt_identity()['role']
+    if user_role == 'ADMIN' or user_role == 'ANALYST':
+        if addItem(request.json['date'], request.json['file']):
+            return jsonify({'msg': 'success'}), 200
+        return jsonify({'msg': 'Error'}), 400
+    return jsonify({'msg': 'Method not allowed'}), 403
 
 @app.route("/api/dev-by/<id>", methods=["GET"])
 @jwt_required(fresh=False)
@@ -92,45 +95,65 @@ def getResource(id):
 @app.route("/api/dev-by/<id>", methods=["DELETE"])
 @jwt_required(fresh=False)
 def deleteResource(id):
-    deleteItem(id)
-    return jsonify({'msg': 'success'}), 200
+    user_role = get_jwt_identity()['role']
+    if user_role == 'ADMIN' or user_role == 'ANALYST':
+        if deleteItem(id) > 0:
+            return jsonify({'msg': 'success'}), 200
+        return jsonify({'msg': 'Error'}), 400
+    return jsonify({'msg': 'Method not allowed'}), 403
+
 
 @app.route("/api/dev-by/parse", methods=["GET"])
 @jwt_required(fresh=False)
 def parseResource():
-    return {'date': str(date.today()), 'records': devbyParse()}
+    user_role = get_jwt_identity()['role']
+    if user_role == 'ADMIN' or user_role == 'ANALYST':
+        return {'date': str(date.today()), 'records': devbyParse()}
+    return jsonify({'msg': 'Method not allowed'}), 403
 
 @app.route("/api/users", methods=["GET"])
 @jwt_required(fresh=False)
 def getUsers():
-    cursor = getAllUsers()
-    data = [record for record in cursor]
+    user_role = get_jwt_identity()['role']
+    if user_role == 'ADMIN':
+        cursor = getAllUsers()
+        data = [record for record in cursor]
 
-    arr = []
+        arr = []
 
-    for item in data:
-        arr.append({
-            '_id': item['_id'],
-            'username': item['username'],
-            'role': item['role'],
-        })
-    arr.reverse()
+        for item in data:
+            arr.append({
+                '_id': item['_id'],
+                'username': item['username'],
+                'role': item['role'],
+            })
+        arr.reverse()
 
-    return arr
+        return arr
+    return jsonify({'msg': 'Method not allowed'}), 403
 
 @app.route("/api/users/<id>", methods=["DELETE"])
 @jwt_required(fresh=False)
 def deleteUser(id):
-    if deleteUserDB(id) > 0:
-        return jsonify({'msg': 'success'}), 200
-    return jsonify({'msg': 'Error'}), 400
+    user_role = get_jwt_identity()['role']
+    if user_role == 'ADMIN':
+        if deleteUserDB(id) > 0:
+            return jsonify({'msg': 'success'}), 200
+        return jsonify({'msg': 'Error'}), 400
+    return jsonify({'msg': 'Method not allowed'}), 403
 
 @app.route("/api/users/<id>", methods=["PUT"])
 @jwt_required(fresh=False)
 def updateUserRole(id):
-    if updateUserRoleDB(id, request.json['role']):
-        return jsonify({'msg': 'success'}), 200
-    return jsonify({'msg': 'Error'}), 400
+    user_role = get_jwt_identity()['role']
+    if user_role == 'ADMIN':
+        if updateUserRoleDB(id, request.json['role']) > 0:
+            return jsonify({'msg': 'success'}), 200
+        return jsonify({'msg': 'Error'}), 400
+    return jsonify({'msg': 'Method not allowed'}), 403
+
+
+CORS(app)
 
 if __name__ == "__main__":
     app.run(debug=True, port=4000)
